requirements.txt
streamlit
wikipedia
google.generativeai
